pyproject.toml
[tool.poetry]
name = "ecom-task1"
version = "0.1.0"
description = ""
authors = ["Michael Rodionov <rodionovspb9618@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.0"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.20"
clickhouse-driver = "^0.2.6"
clickhouse-connect = "^0.6.11"
pydantic-settings = "^2.0.3"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
